Refactor multi-thread julia support into its own module
