+++ src/share/Goby.jl/src/Goby.jl
@@ -3,7 +3,6 @@ module Goby
 # Load the module and generate the functions
 using CxxWrap
 using ProtoBuf
-using ThreadPools
 
 export publish, subscribe
 
@@ -24,17 +23,20 @@ cxx_task_id = TaskID(1)
 
 is_multithreaded = false
 
+include("GobyMultiThread.jl")
+
 # main task Protobuf publish
 # TODO: add more schemes as additional publish functions
 function publish(app, layer, group, msg::AbstractProtoBufMessage)
-    layer_int::Int32 = Int32(layer)    
+    layer_int::Int32 = Int32(layer)
+    
     if layer_int == Int32(Goby.INTERTHREAD)
-        publish_interthread(main_task_id, group, msg)
+        MultiThread.publish_interthread(main_task_id, group, msg)
         return
     end
     
     if Goby.is_multithreaded && Threads.threadid() != Goby.cxx_task_id.id
-        publish_forward_interprocess(app, layer, group, msg)
+        MultiThread.publish_forward_interprocess(app, layer, group, msg)
         return
     end
     
@@ -58,7 +60,7 @@ function publish(app, layer, group, msg)
     layer_int::Int32 = Int32(layer)
     if layer_int == Int32(Goby.INTERTHREAD)
         # this would be the main task as all child tasks would use task_id::TaskID method
-        publish_interthread(main_task_id, group, msg)
+        MultiThread.publish_interthread(main_task_id, group, msg)
         return
     end
 
@@ -70,7 +72,7 @@ end
 function publish(task_id::TaskID, layer, group, msg)
     layer_int::Int32 = Int32(layer)
     if layer_int == Int32(Goby.INTERTHREAD)
-        publish_interthread(task_id, group, msg)
+        MultiThread.publish_interthread(task_id, group, msg)
         return
     end
 
@@ -91,7 +93,7 @@ end
 function subscribe(task_id::TaskID, layer, group, callback::Function)
     layer_int::Int32 = Int32(layer)
     if layer_int == Int32(Goby.INTERTHREAD)        
-        subscribe_interthread(task_id, group, callback)
+        MultiThread.subscribe_interthread(task_id, group, callback)
         return
     end
 
@@ -103,7 +105,7 @@ function subscribe(app, layer, group, callback::Function; scheme = Goby.NULL_SCH
     layer_int::Int32 = Int32(layer)
 
     if layer_int == Int32(Goby.INTERTHREAD)        
-        subscribe_interthread(main_task_id, group, callback)
+        MultiThread.subscribe_interthread(main_task_id, group, callback)
         return
     end
 
@@ -158,7 +160,7 @@ function receive(cxx_layer, cxx_type_name, cxx_scheme, cxx_group, vec::CxxRef{St
 
 
     if Goby.is_multithreaded && Threads.threadid() != Goby.main_task_id.id
-        receive_forward_interprocess(layer, type_name, scheme, group, bytes)
+        MultiThread.receive_forward_interprocess(layer, type_name, scheme, group, bytes)
         return
     end
 
@@ -193,200 +195,20 @@ function read_cli_cfg()
 end
 
 function run(goby_app)
-    if isdefined(Main, :cfg) && haskey(Main.cfg, :loop_frequency)
-        Goby.cxx_set_loop_frequency_hertz(goby_app, Main.cfg[:loop_frequency])
+    if isdefined(Main, :goby_cfg) && haskey(Main.goby_cfg, :loop_frequency)
+        Goby.cxx_set_loop_frequency_hertz(goby_app, Main.goby_cfg[:loop_frequency])
     end
     Goby.cxx_run(goby_app)
 end
 
 function cxx_loop()
     if Goby.is_multithreaded
-        # Julia loop is called from timer trigger
-        # but we use C++ loop() to let cxx_task check its channel
-        while isready(Goby.task_interthread_channels[Goby.cxx_task_id])
-            Goby.task_channel_check(Goby.cxx_task_id)
-        end
+        MultiThread.cxx_loop()
     else
         # Julia loop is directly called from C++ loop()
-        Main.cfg[:loop_function]()
+        Main.goby_cfg[:loop_function]()
     end
 end
 
-########################
-# Interthread          #
-# Using Tasks/Channels #
-########################
-
-# map of module to task_id
-task_module_to_id=Dict{Module, TaskID}()
-# map of task_id to vector of channels (used for 'check')
-task_interthread_channels=Dict{TaskID, Channel}()
-channel_size = 50
-task_interthread_channels[main_task_id] = Channel(channel_size)
-
-# map of group to vector of channels (used for 'publish')
-group_interthread_channels=Dict{String, Vector{Channel}}()
-group_interthread_channels_lock = Threads.ReentrantLock()
-
-function publish_interthread(task_id::TaskID, group, msg)
-    Threads.lock(group_interthread_channels_lock) do
-        if haskey(group_interthread_channels, group)
-            for channel in group_interthread_channels[group]
-                put!(channel, (:interthread_receive, group, msg))
-            end
-        end
-    end
-end
-
-function publish_forward_interprocess(app, layer, group, msg)
-    put!(task_interthread_channels[Goby.cxx_task_id], (:cxx_publish, app, layer, group, msg))
-end
-        
-function receive_forward_interprocess(layer, type_name, scheme, group, bytes)
-    put!(task_interthread_channels[Goby.main_task_id], (:cxx_receive, layer, type_name, scheme, group, bytes))
-end
-
-
-function subscribe_interthread(task_id::TaskID, group, callback::Function)
-    Threads.lock(group_interthread_channels_lock) do
-        if !haskey(Threads.task_local_storage(), :callbacks)
-            task_local_storage(:callbacks, Dict{String, Vector{Function}}())
-        end
-        task_callbacks = task_local_storage(:callbacks)
-        
-        group_callbacks = get!(task_callbacks, group) do
-            Vector{Function}()
-        end
-        
-        push!(group_callbacks, callback)
-        
-        group_interthread_channel = get!(group_interthread_channels, group) do
-            Vector{Channel}()
-        end
-
-        push!(group_interthread_channel, task_interthread_channels[task_id])
-    end
-            
-    println("Subscribed $task_id to $group")
-end
-    
-function run(goby_app, main_module, task_modules)
-    tasks = Vector{Task}()
-    Goby.is_multithreaded = true
-
-    # 1. set cxx loop frequency, defaulting to 10 Hz (used for checking channel messages)
-    cxx_channel_check_frequency = 10
-    if isdefined(Main, :cfg) && haskey(Main.cfg, :cxx_channel_check_frequency)
-        cxx_channel_check_frequency = Main.cfg[:cxx_channel_check_frequency]
-    end
-    println("Setting cxx channel check frequency to $cxx_channel_check_frequency")
-    Goby.cxx_set_loop_frequency_hertz(goby_app, cxx_channel_check_frequency)
-    
-    # 2. create channels
-    for task_module in task_modules
-        task_id = TaskID(Goby.next_task_id.id)
-        task_module_to_id[task_module] = task_id
-        task_interthread_channels[task_id] = Channel(Goby.channel_size)
-        Goby.next_task_id = TaskID(Goby.next_task_id.id + 1)
-    end
-    
-    task_module_to_id[main_module] = main_task_id
-    Goby.cxx_task_id = TaskID(Goby.next_task_id.id)        
-    task_interthread_channels[Goby.cxx_task_id] = Channel(Goby.channel_size)
-    
-    
-    # 3. spawn tasks
-    for task_module in task_modules
-        push!(tasks, Goby.task_spawn(task_module))
-        
-        if haskey(task_module.cfg, :loop_function)
-            push!(tasks, Goby.loop_timer(task_module))
-        end
-    end
-    
-    push!(tasks, ThreadPools.@tspawnat cxx_task_id.id Goby.cxx_run(goby_app))
-    println("Spawning main run() as task ID $cxx_task_id")
-   
-    if haskey(main_module.cfg, :loop_function)
-        push!(tasks, Goby.loop_timer(main_module))
-    end
-    
-    # 4. Loop over all tasks checking messages
-    while true
-        Goby.task_channel_check(main_task_id)
-        for task in tasks
-            if istaskdone(task)
-                fetch(task)
-            end
-        end
-    end
-end
-
-
-function task_spawn(task_module::Module)
-    task_id = task_module_to_id[task_module]
-    println("Spawning $run as task ID $task_id")
-
-    runner = (id, mod::Module) -> begin
-        mod.run(id)
-        while true
-            Goby.task_channel_check(id)
-        end
-    end
-    
-    t = ThreadPools.@tspawnat task_id.id runner(task_id, task_module)
-    println("Task: $t")
-    return t
-end
-
-
-function loop_timer(task_module::Module)
-    task_id = task_module_to_id[task_module]
-
-    timer = (id, mod::Module) -> begin
-        while true
-            put!(task_interthread_channels[id], (:loop, mod.cfg[:loop_function]))
-            sleep(1/mod.cfg[:loop_frequency])
-#            println("Loop timer: $task_id")
-        end
-    end
-
-    t = ThreadPools.@tspawnat timer_task_id.id timer(task_id, task_module)
-    return t
-end    
-
-function task_channel_check(task_id::TaskID)
-    packet = take!(task_interthread_channels[task_id])
-
-#    println("Task $task_id received packet $packet")
-    type = packet[1]
-    if type == :interthread_receive
-        task_callbacks = task_local_storage(:callbacks)
-        group = packet[2]
-        msg = packet[3]
-        if haskey(task_callbacks, group)
-            callbacks = task_callbacks[group]
-            for callback in callbacks
-                callback(task_id, msg)
-            end
-        end
-    elseif type == :loop
-        loop_func = packet[2]
-        loop_func(task_id)
-    elseif type == :cxx_publish
-        app = packet[2]
-        layer = packet[3]
-        group = packet[4]
-        msg = packet[5]
-        publish(app, layer, group, msg)
-    elseif type == :cxx_receive
-        layer = packet[2]
-        type_name = packet[3]
-        scheme = packet[4]
-        group = packet[5]
-        bytes = packet[6]
-        receive_dereferenced(layer, type_name, scheme, group, bytes)
-    end    
-end
 
 end # module Goby
